add router create_new_product

+++ pizzeria_app/app/dto/order_details/schema.py
@@ -1,5 +1,6 @@
 from uuid import UUID
 
+from decimal import Decimal
 from pydantic import BaseModel
 
 
@@ -7,7 +8,7 @@ class OrderDetails(BaseModel):
     order_id: UUID
     product_id: UUID
     quantity: int
-    unit_price: int
+    unit_price: Decimal
 
     class Config:
         from_attributes = True
